requirements.txt
torch>=1.10
torchvision>=0.11
numpy
matplotlib
scikit-image
tqdm
Pillow
lpips
torchmetrics
